VerifyBackend/requirements.txt
fastapi
uvicorn
torch
numpy
pillow
python-dotenv
supabase
